Refactor the LMS code.
The previous version was very specific and contained duplicate code.
The current version scrambles Joe&Kuo matrices, but could easily be adapted for other matrices.
It remains to be tested.
Also the CLI interface should be improved, in relation with the change of the "SizeParameter" option which is not a really good name.