Change the polynomial I/O from bit arrays to ints
Previously, the I/O of polynomials was arrays of bits, where the last bit corresponds to the leading coefficient (the same as in NTL).
This commit changes the I/O to integers, the standard format in the QMC community, where each monomial a_iX^i is represented by the integer a_i 2^i.
In particular, the bit order is reversed between both representations.
Some redundant functions (polynomial to int conversion, int exponentiation) are removed.